# internal/docs/swagger.yaml
basePath: /api
definitions:
  account.Provider:
    enum:
    - email
    - google
    - github
    - email
    - apple
    type: string
    x-enum-varnames:
    - DefaultProvider
    - ProviderGoogle
    - ProviderGithub
    - ProviderEmail
    - ProviderApple
  controller.ErrorResponse:
    properties:
      error: {}
      status:
        type: integer
    type: object
  data.LoginBody:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  data.LoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/ent.User'
    type: object
  data.UserCreateBody:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - email
    - name
    - password
    type: object
  ent.Account:
    properties:
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/ent.AccountEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the AccountQuery when eager-loading is set.
      email:
        description: Email holds the value of the "email" field.
        type: string
      id:
        description: ID of the ent.
        type: string
      mail_verified:
        description: MailVerified holds the value of the "mail_verified" field.
        type: boolean
      provider:
        allOf:
        - $ref: '#/definitions/account.Provider'
        description: Provider holds the value of the "provider" field.
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
      user_id:
        description: UserID holds the value of the "user_id" field.
        type: string
    type: object
  ent.AccountEdges:
    properties:
      sessions:
        description: Sessions holds the value of the sessions edge.
        items:
          $ref: '#/definitions/ent.Session'
        type: array
      user:
        allOf:
        - $ref: '#/definitions/ent.User'
        description: User holds the value of the user edge.
    type: object
  ent.Session:
    properties:
      account_id:
        description: AccountID holds the value of the "account_id" field.
        type: string
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/ent.SessionEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the SessionQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: string
      ip_address:
        description: IPAddress holds the value of the "ip_address" field.
        type: string
      token:
        description: Token holds the value of the "token" field.
        type: string
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
      user_agent:
        description: UserAgent holds the value of the "user_agent" field.
        type: string
    type: object
  ent.SessionEdges:
    properties:
      owner:
        allOf:
        - $ref: '#/definitions/ent.Account'
        description: Owner holds the value of the owner edge.
    type: object
  ent.User:
    properties:
      Name:
        description: Name holds the value of the "Name" field.
        type: string
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/ent.UserEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the UserQuery when eager-loading is set.
      email:
        description: Email holds the value of the "email" field.
        type: string
      id:
        description: ID of the ent.
        type: string
      image_url:
        description: ImageURL holds the value of the "image_url" field.
        type: string
      role:
        allOf:
        - $ref: '#/definitions/user.Role'
        description: Role holds the value of the "role" field.
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
    type: object
  ent.UserEdges:
    properties:
      account:
        allOf:
        - $ref: '#/definitions/ent.Account'
        description: Account holds the value of the account edge.
    type: object
  user.Role:
    enum:
    - user
    - admin
    - user
    type: string
    x-enum-varnames:
    - DefaultRole
    - RoleAdmin
    - RoleUser
info:
  contact: {}
  description: API for UpChat - AI chat application
  title: UpChat API
  version: "1.0"
paths:
  /auth/get-session:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.User'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current session
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/data.LoginBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/data.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Log in
      tags:
      - auth
  /auth/logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Log out
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      parameters:
      - description: User registration data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/data.UserCreateBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: SignUp
      tags:
      - auth
  /docs/openapi.json:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: OpenAPI spec
      tags:
      - docs
  /health:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check
      tags:
      - health
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"